backend/requirements.txt
# ==========================================
# Core Dependencies
# ==========================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ==========================================
# Authentication & Security
# ==========================================
PyJWT==2.8.0
email-validator==2.1.0

# ==========================================
# Database & Caching
# ==========================================
redis==5.0.1
aioredis==2.0.1
SQLAlchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ==========================================
# AI/ML Providers
# ==========================================
openai==1.6.1
anthropic==0.8.1

# NLP & Embeddings
sentence-transformers==2.2.2
tiktoken==0.5.2

# ==========================================
# Data Processing
# ==========================================
numpy==1.26.3
pandas==2.1.4
PyYAML==6.0.1

# ==========================================
# HTTP & Networking
# ==========================================
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# ==========================================
# Monitoring & Observability
# ==========================================
prometheus-client==0.19.0
psutil==5.9.6

# OpenTelemetry
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-exporter-jaeger==1.21.0

# ==========================================
# File Processing & Utilities
# ==========================================
python-multipart==0.0.6
pillow==10.1.0
beautifulsoup4==4.12.2

# ==========================================
# Cloud & Infrastructure
# ==========================================
boto3==1.34.14
kubernetes==29.0.0

# ==========================================
# Task Scheduling
# ==========================================
schedule==1.2.0

# ==========================================
# Dashboard & Visualization (FinOps)
# ==========================================
streamlit==1.29.0
plotly==5.18.0

# ==========================================
# Development & Testing
# ==========================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0
locust==2.17.0
coverage==7.3.2

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
pre-commit==3.6.0